Moved team member functions to Team Member files
