+++ src/main/resources/db/migration/V11__add_profile_to_users.sql
@@ -1,0 +1,2 @@
+ALTER TABLE users
+ADD COLUMN profile VARCHAR(255);
